Fix tpm*_id failing to find existing NVRAM area
Do not use `-q` flag of grep.  The flag makes grep exit on finding a
match before reaching EOF.  Closing of grep's input stream can result in
tpm_nvinfo/tpm2_nvreadpublic failing to write to its output stream and
error, which then becomes the result of the whole pipe due to the use of
`set -o pipefail` at the top.

Observed on Qubes OS 4.2 with TPM2, but updated tpm_id as well just in
case it's affected or will be affected in the future.  This wasn't
caught before probably because the behaviour has changed either in grep
or in tpm2_nvreadpublic.

Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>